Revert "Changed the order of games, and changed a few colors"
