+++ package.json
@@ -12,25 +12,22 @@
     "dependencies": {
         "@google-cloud/speech": "^6.7.0",
         "@google-cloud/text-to-speech": "^5.7.0",
-        "@prisma/client": "^6.5.0",
+        "@prisma/client": "^6.3.1",
         "body-parser": "^1.20.3",
         "boxicons": "^2.1.4",
         "cors": "^2.8.5",
         "dotenv": "^16.4.7",
         "express": "^4.21.2",
-        "jwt-decode": "^4.0.0",
         "vercel": "^39.2.5",
         "vue": "^3.4.31",
-        "vue-router": "^4.4.0",
-        "vue3-google-login": "^2.0.33",
-        "vue3-google-signin": "^2.1.1"
+        "vue-router": "^4.4.0"
     },
     "devDependencies": {
         "@vitejs/plugin-vue": "^5.0.5",
         "autoprefixer": "^10.4.19",
         "gh-pages": "^6.2.0",
         "postcss": "^8.4.40",
-        "prisma": "^6.5.0",
+        "prisma": "^6.3.1",
         "sass-embedded": "^1.83.1",
         "tailwindcss": "^3.4.7",
         "vite": "^5.3.4"
