+++ src/pages/LogIn/LogIn.vue
@@ -1,131 +1,52 @@
 <script setup>
-import { ref, onMounted } from "vue";
-import { GoogleLogin } from "vue3-google-login";
-import { useRouter } from "vue-router";
-import { jwtDecode } from "jwt-decode";
+import { ref } from "vue";
 
 const errors = ref(false);
 const email = ref("");
 const password = ref("");
-var authKey = ref("");
-const userSession = ref(null);
-const userProfile = ref(null);
-const school = ref(""); // Store school input
-const showSchoolForm = ref(false); // Control form visibility
-const router = useRouter();
+let authKey = ref("");
 
-onMounted(() => {
-    const session = localStorage.getItem("audemyUserSession");
-    if (session) {
-        userSession.value = JSON.parse(session);
-    }
-});
-
-const login = async (event) => {
+const login = (event) => {
     event.preventDefault();
     errors.value = false;
+
     if (!email.value || !password.value) {
         errors.value = true;
         resetErrors();
         return;
     }
 
-    try {
-        const response = await fetch("https://audemy-users-api.fly.dev/login", {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({
-                user: { email: email.value, password: password.value },
-            }),
+    // Send login req data as JSON
+    fetch("https://audemy-users-api.fly.dev/login", {
+        method: "POST",
+        headers: {
+            "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+            user: {
+                email: email.value,
+                password: password.value,
+            },
+        }),
+    })
+        .then((response) => {
+            response.json();
+            //This is where the auth key is saved for future API calls
+            //We should think about saving it in a more central location using Vuex library
+            authKey = response.headers.get("authorization");
+        })
+        .then((data) => {
+            console.log("Success:", data);
+        })
+        .catch((error) => {
+            console.error("Error:", error);
         });
-
-        const data = await response.json();
-        authKey.value = response.headers.get("authorization");
-
-        localStorage.setItem(
-            "audemyUserSession",
-            JSON.stringify({ token: authKey.value, user: data.user })
-        );
-        userSession.value = { token: authKey.value, user: data.user };
-    } catch (error) {
-        console.error("Error:", error);
-    }
 };
 
-const callback = async (response) => {
-    if (response?.credential) {
-        try {
-            const decoded = jwtDecode(response.credential);
-            userProfile.value = {
-                name: decoded.name,
-                email: decoded.email,
-                imageUrl: decoded.picture,
-            };
-        } catch (error) {
-            console.error("Failed to decode JWT:", error);
-        }
-    }
-
-    const dbResponse = await fetch(
-        `/api/db/get_user?email=${userProfile.value.email}`,
-        {
-            method: "GET",
-            headers: { "Content-Type": "application/json" },
-        }
-    );
-
-    const dbData = await dbResponse.json();
-    console.log("DB Response:", dbData);
-
-    if (!dbData || !dbData.email) {
-        console.log("User not found, prompting for school...");
-        showSchoolForm.value = true;
-    } else {
-        localStorage.setItem("audemyUserSession", JSON.stringify(response));
-        userSession.value = response;
-        router.push("/game-zone");
-    }
-};
-
-const updateSchool = async () => {
-    if (!school.value) {
-        alert("Please enter your school name.");
-        return;
-    }
-
-    try {
-        const response = await fetch(`/api/db/update_user_school`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({
-                email: userProfile.value.email,
-                name: userProfile.value.name,
-                school: school.value,
-            }),
-        });
-
-        const data = await response.json();
-        console.log("Updated user:", data);
-
-        if (data.success) {
-            localStorage.setItem(
-                "audemyUserSession",
-                JSON.stringify(userProfile.value)
-            );
-            userSession.value = userProfile.value;
-            showSchoolForm.value = false;
-            router.push("/game-zone");
-        }
-    } catch (error) {
-        console.error("Error updating school:", error);
-    }
-};
-
-const logout = () => {
-    localStorage.removeItem("audemyUserSession");
-    userSession.value = null;
-    userProfile.value = null;
-    router.push("/login");
+const resetErrors = () => {
+    setTimeout(() => {
+        errors.value = false;
+    }, 4000);
 };
 </script>
 
@@ -155,24 +76,15 @@ const logout = () => {
                 alt="wave icon"
                 class="absolute -bottom-[15%] right-0 w-full -z-1"
             />
-            <button
-                v-if="userSession"
-                @click="logout"
-                class="mt-4 bg-red-500 text-white px-4 py-2 rounded"
-            >
-                Logout
-            </button>
         </div>
 
-        <!-- Show login form if not logged in -->
         <div
-            v-if="!userSession && !showSchoolForm"
-            class="w-7/12 md:w-full sm:w-full bg-white flex flex-col items-center justify-center border-2"
+            class="w-7/12 md:w-full sm:w-full bg-white flex items-center justify-center"
         >
             <form
                 @submit="login"
                 method="post"
-                class="max-h-[350px] w-full flex flex-col justify-center items-center gap-[5%] my-4"
+                class="max-h-[350px] w-full flex flex-col justify-center items-center gap-[5%]"
             >
                 <h1
                     class="text-[36px] text-[#151E22] text-center w-7/12 mobile:w-full mobile:text-[24px] mobile:mb-4"
@@ -191,12 +103,14 @@ const logout = () => {
                             >
                         </div>
                     </div>
+                    <!-- Changed for field here to email -->
                     <div class="mb-[16px]">
                         <label
                             class="block text-[#0C0D0D] mb-1 font-semiBold"
                             for="email"
                             >Email Address</label
                         >
+                        <!-- Added Autocomplete here -->
                         <input
                             v-model="email"
                             type="email"
@@ -222,12 +136,16 @@ const logout = () => {
                             placeholder="Enter your password"
                         />
                     </div>
-                    <div class="flex justify-end w-full">
-                        <a href="#" class="underline text-[#087BB4] font-medium"
+                    <div
+                        class="flex justify-end w-full mobile:items-center mobile:text-center mobile:justify-center"
+                    >
+                        <a href="" class="underline text-[#087BB4] font-medium"
                             >Forgot password?</a
                         >
                     </div>
-                    <div class="flex justify-start w-full">
+                    <div
+                        class="flex justify-start w-full mobile:items-center mobile:text-center mobile:justify-center mobile:pt-4"
+                    >
                         <h5>
                             New to Audemy?
                             <a
@@ -237,48 +155,22 @@ const logout = () => {
                             >
                         </h5>
                     </div>
-
                     <div class="flex justify-center w-full pt-4">
                         <button
                             type="submit"
-                            class="w-full py-3 font-bold rounded-[8px] bg-[#FE892A] hover:bg-[#ff8d33] border-2 border-black shadow-[4px_4px_0px_black] text-black"
+                            class="w-full py-3 font-bold rounded-[8px]"
+                            :class="
+                                errors
+                                    ? 'bg-[#747575] text-white'
+                                    : 'bg-[#FE892A] hover:bg-[#ff8d33] border-2 border-black shadow-[4px_4px_0px_black] text-black'
+                            "
+                            :disabled="errors"
                         >
                             Log in
                         </button>
                     </div>
                 </div>
             </form>
-
-            <div class="flex w-1/2 text-gray-500 items-center justify-center gap-2">
-                <div><hr class="w-52 h-0.5 my-4 bg-gray-500 rounded-sm"/></div>
-                <div>or</div>
-                <div><hr class="w-52 h-0.5 my-4 bg-gray-500 rounded-sm"/></div>
-            </div>
-
-            <!-- Google OAuth Login -->
-            <div class="flex w-full gap-4 items-center justify-center mt-4">
-                <GoogleLogin :callback="callback" class=" flex items-center justify-center gap-4"/>
-            </div>
-        </div>
-
-        <!-- Show school input if needed -->
-        <div
-            v-if="showSchoolForm"
-            class="w-7/12 bg-white flex items-center justify-center flex-col gap-4"
-        >
-            <h3>Enter Your School</h3>
-            <input
-                v-model="school"
-                type="text"
-                class="border-2 border-black py-2 px-2 rounded-[8px]"
-                placeholder="School Name"
-            />
-            <button
-                @click="updateSchool"
-                class="bg-blue-500 text-white py-2 px-4 rounded mt-4"
-            >
-                Submit
-            </button>
         </div>
     </div>
 </template>
